move the get_dtype_from_torch to the cpp file prevent redefination
